Split out the z3 stuff from the more solid disease stuff
