+++ doc/src/README.md
@@ -22,18 +22,19 @@ The assembler lives in the [`etk-asm`] crate, and provides a command-line interf
 
 ### Disassembler: `disease`
 
-The disassembler lives in the [`etk-analyze`] crate, and also provides a command-line interface (`disease`) and a Rust library. The disassembler is much more experimental, and has a limited set of features:
+The disassembler lives in the [`etk-dasm`] crate, and also provides a command-line interface (`disease`) and a Rust library. The disassembler is much more experimental, and has a limited set of features:
 
  - Disassemble hex or binary encoded instructions into their mnemonics.
  - Identify basic blocks.
 
-[`etk-analyze`]: https://crates.io/crates/etk-analyze
+[`etk-dasm`]: https://crates.io/crates/etk-dasm
 
 ## API Documentation
 
 Alongside this book, you can also read the API docs generated by Rustdoc if you would like to use ETK as a library.
 
  - [etk-asm](https://docs.rs/etk-asm)
+ - [etk-dasm](https://docs.rs/etk-dasm)
  - [etk-analyze](https://docs.rs/etk-analyze)
 
 ### Dependencies
