+++ etk-analyze/src/blocks.rs
@@ -1,7 +1,1 @@
-//! Blocks are sections of EVM instructions.
-
-pub mod annotated;
-pub mod basic;
-
-pub use self::annotated::AnnotatedBlock;
-pub use self::basic::BasicBlock;
+pub(crate) mod annotated;
