+++ etk-dasm/README.md
@@ -1,0 +1,9 @@
+etk-dasm
+========
+
+The EVM Toolkit's disassembler: `disease`.
+
+## Documentation
+
+ - [The ETK Book (master branch)](https://quilt.github.io/etk/)
+ - [API Documentation](https://docs.rs/etk-dasm)
